reorganize repo to accomodate overrides for legacy torch version support
