+++ .pre-commit-config.yaml
@@ -1,19 +1,28 @@
 repos:
--   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v2.3.0
-    hooks:
-    -   id: check-yaml
-    -   id: end-of-file-fixer
-    -   id: trailing-whitespace
--   repo: local
-    hooks:
-    -   id: black
-        name: black
-        entry: python -m black
-        language: system
-        types: [python]
-    -   id: flake8
-        name: flake8
-        entry: python -m flake8
-        language: system
-        types: [python]
+- repo: https://github.com/pre-commit/pre-commit-hooks
+  rev: v3.3.0
+  hooks:
+  -   id: check-yaml
+  -   id: end-of-file-fixer
+  -   id: trailing-whitespace
+- repo: git@github.com:humitos/mirrors-autoflake.git
+  rev: v1.1
+  hooks:
+  - id: autoflake
+    name: autoflake
+    entry: python -m autoflake
+    language: system
+    args: ['--in-place', '--remove-all-unused-imports', '--remove-unused-variable']
+
+- repo: local
+  hooks:
+  -   id: black
+      name: black
+      entry: python -m black
+      language: system
+      types: [python]
+  -   id: flake8
+      name: flake8
+      entry: python -m flake8
+      language: system
+      types: [python]
