+++ configen/conf/configen.yaml
@@ -38,7 +38,7 @@ configen:
     - name: torch.optim.sparse_adam
       # for each module, a list of classes
       classes:
-        - SparseAdam 
+        - SparseAdam
 
     - name: torch.optim.adamax
       # for each module, a list of classes
@@ -48,27 +48,27 @@ configen:
     - name: torch.optim.asgd
       # for each module, a list of classes
       classes:
-        - ASGD 
+        - ASGD
 
     - name: torch.optim.sgd
       # for each module, a list of classes
       classes:
-        - SGD 
+        - SGD
 
     - name: torch.optim.rprop
       # for each module, a list of classes
       classes:
-        - Rprop 
+        - Rprop
 
     - name: torch.optim.rmsprop
       # for each module, a list of classes
       classes:
-        - RMSprop 
+        - RMSprop
 
     - name: torch.optim.lbfgs
       # for each module, a list of classes
       classes:
-        - LBFGS 
+        - LBFGS
 
     - name: torch.optim.lr_scheduler
       classes:
