+++ hydra_configs/torch/optim/__init__.py
@@ -1,0 +1,16 @@
+# This file mirrors the __init__.py within pytorch/torch/optim
+# It allows hydra-torch configs to be imported in the same style
+# flake8: noqa
+
+
+del adadelta
+del adagrad
+del adam
+del adamw
+del sparse_adam
+del adamax
+del asgd
+del sgd
+del rprop
+del rmsprop
+del lbfgs
