+++ hydra_configs/torch/optim/adadelta.py
@@ -6,17 +6,16 @@
 # isort:skip_file
 # flake8: noqa
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from omegaconf import MISSING
 from typing import Any
 
 
 @dataclass
-class ASGDConf:
-    _target_: str = "torch.optim.asgd.ASGD"
+class AdadeltaConf:
+    _target_: str = "torch.optim.adadelta.Adadelta"
     params: Any = MISSING
-    lr: Any = 0.01
-    lambd: Any = 0.0001
-    alpha: Any = 0.75
-    t0: Any = 1000000.0
+    lr: Any = 1.0
+    rho: Any = 0.9
+    eps: Any = 1e-06
     weight_decay: Any = 0
