+++ hydra_configs/torch/optim/adam.py
@@ -6,16 +6,17 @@
 # isort:skip_file
 # flake8: noqa
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from omegaconf import MISSING
 from typing import Any
 
 
 @dataclass
-class AdamaxConf:
-    _target_: str = "torch.optim.adamax.Adamax"
+class AdamConf:
+    _target_: str = "torch.optim.adam.Adam"
     params: Any = MISSING
-    lr: Any = 0.002
+    lr: Any = 0.001
     betas: Any = (0.9, 0.999)
     eps: Any = 1e-08
     weight_decay: Any = 0
+    amsgrad: Any = False
