+++ hydra_configs/torch/optim/adamax.py
@@ -6,17 +6,16 @@
 # isort:skip_file
 # flake8: noqa
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from omegaconf import MISSING
 from typing import Any
 
 
 @dataclass
-class AdamConf:
-    _target_: str = "torch.optim.adam.Adam"
+class AdamaxConf:
+    _target_: str = "torch.optim.adamax.Adamax"
     params: Any = MISSING
-    lr: Any = 0.001
+    lr: Any = 0.002
     betas: Any = (0.9, 0.999)
     eps: Any = 1e-08
     weight_decay: Any = 0
-    amsgrad: Any = False
