+++ hydra_configs/torch/optim/sparse_adam.py
@@ -6,18 +6,15 @@
 # isort:skip_file
 # flake8: noqa
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from omegaconf import MISSING
 from typing import Any
 
 
 @dataclass
-class RMSpropConf:
-    _target_: str = "torch.optim.rmsprop.RMSprop"
+class SparseAdamConf:
+    _target_: str = "torch.optim.sparse_adam.SparseAdam"
     params: Any = MISSING
-    lr: Any = 0.01
-    alpha: Any = 0.99
+    lr: Any = 0.001
+    betas: Any = (0.9, 0.999)
     eps: Any = 1e-08
-    weight_decay: Any = 0
-    momentum: Any = 0
-    centered: Any = False
